+++ BOCCHI.World/Services/FateContext.cs
@@ -1,4 +1,5 @@
 ﻿using BOCCHI.Common.Data.Fates;
+using BOCCHI.Common.Extensions;
 using BOCCHI.Common.Services;
 using Dalamud.Game.ClientState.Objects.SubKinds;
 using Dalamud.Game.ClientState.Objects.Types;
@@ -40,15 +41,49 @@ public class FateContext(IObjectTable objects) : IFateContext
         return objects.OfType<IBattleNpc>()
             .Where(obj => obj is { IsDead: false, IsTargetable: true })
             .Where(o => o.IsHostile())
-            .Where(obj =>
-            {
-                unsafe
-                {
-                    BattleChara* battleChara = (BattleChara*)obj.Address;
-
-                    return battleChara->FateId == fateId;
-                }
-            })
+            .Where(obj => NpcBelongsToFate(obj, fateId))
             .OrderBy(o => o.Position.Distance2D(player.Position));
+    }
+
+    public bool IsInCombatWith(FateId id)
+    {
+        IPlayerCharacter? player = objects.LocalPlayer;
+        if (player == null)
+        {
+            return false;
+        }
+
+        ulong? targetId = player.TargetObject?.GameObjectId;
+
+        foreach (IBattleNpc npc in objects.OfType<IBattleNpc>())
+        {
+            if (npc is not { IsDead: false, IsTargetable: true } || !npc.IsHostile())
+            {
+                continue;
+            }
+
+            if (!NpcBelongsToFate(npc, id.Value))
+            {
+                continue;
+            }
+
+            if (targetId is { } idValue && npc.GameObjectId == idValue)
+            {
+                return true;
+            }
+
+            if (npc.IsTargetingPlayer(player))
+            {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    private static unsafe bool NpcBelongsToFate(IBattleNpc npc, ushort fateId)
+    {
+        BattleChara* battleChara = (BattleChara*)npc.Address;
+        return battleChara->FateId == fateId;
     }
 }